+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for PayPal Checkout for Craft Commerce
 
+## Unreleased
+
+- Added `craft\commerce\paypalcheckout\gateways\Gateway::EVENT_BUILD_GATEWAY_REQUEST`.
+- Added `\craft\commerce\paypalcheckout\events\BuildGatewayRequestEvent`.
+
 ## 3.0.1 - 2023-06-27 [CRITICAL]
 
 - Added `craft\commerce\paypalcheckout\gateways\Gateway::showPaymentFormSubmitButton()`.